Cookie - 클라이언트 URL 명시 (#145)
* Changjun/infra (#136)

* rds(table) : table name to uppercase

* rds(table) : table name to lowercase

* rds(table) : table name to lowercase && entity pk JdbcTypeCode injection

* rename(project) : 컨트롤러, 서비스단 메서드 이름 재정의

* refactor(user) : 메서드 이름 변경, 로직 세분화, 예외 추가

* refactor(project) : 로직 업데이트

* refactor(member) : 로직 업데이트

* refactor(issue) : 이슈 삭제시 Id 반환
이슈 생성 및 수정 Optional 허용

* refactor(issue) : 이슈 컨트롤러 반환 타입 명시

* refactor(issue) : 이슈 검색 조회, 이슈 조회 로직 업데이트,
에러 케이스 추가

* test(user) : user 단위 테스트 완료

* test(member) : 멤버 도메인 단위 테스트

* refactor(issue) : 이슈 생성, 수정 로직 변경, assignee optional 허용

* refactor(comment) : comment 로직 업데이트

* refactor(issue) : 조회 리스폰스 null 허용

* feat(issue) : 검색 필터 관련 validator 추가

* refactor(issue) : 이슈 페이징 처리 관련 로직 업데이트

* refactor(project) : dto에 프로젝트 이미지 추가

* refactor(project) : project img 데이터 처리

* refactor(member) : 사용자 삭제 관리, cors 설정 롤백

* refactor(member) : 사용자 삭제 로직 업데이트

* refactor(project) : 작업 내역 반영

* refactor(member) : RequestSignIn 제약 추가

* 업데이트 내역 반영 (#137) (#138)

* refactor(member) : RequestSignIn 제약 추가

---------

* refactor(member) : 운영 환경 -> 불필요한 채팅 요소 삭제

* refactor(task) : 불필요한 항목 제거

* refactor(chatting) : 사용하지 않는 컬럼 정리 (chat .. comment check..)

* refactor(user) : 로그인 dto 레코드 -> 클래스로 변경

* refactor(user) : 프로필 이미지 변경시 null 처리 업데이트

* refactor(user) : 프로필 이미지 변경시 null 처리 업데이트

* refactor(project) : 프로젝트 조회시 리더 누락 항목 수정

* refactor(user) : 비밀번호 업데이트 트랜잭션 적용, 저장 명시

* refactor(cookie) : 클라이언트 URL 명시

+++ src/main/java/com/task/weaver/domain/member/service/impl/MemberServiceImpl.java
@@ -94,8 +94,12 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public void logout(String refreshToken) {
-        Authentication authentication = jwtTokenProvider.getAuthentication(resolveToken(refreshToken));
-        refreshTokenRepository.deleteById(authentication.getName());
+        if (jwtTokenProvider.validateToken(refreshToken)) {
+            Authentication authentication = jwtTokenProvider.getAuthentication(refreshToken);
+            refreshTokenRepository.deleteById(authentication.getName());
+            return;
+        }
+        throw new CannotResolveToken(REFRESH_TOKEN_RESOLVE, REFRESH_TOKEN_RESOLVE.getMessage());
     }
 
     @Override
@@ -274,9 +278,10 @@ public class MemberServiceImpl implements MemberService {
         }
     }
 
-    private String resolveToken(String accessToken) {
-        if (accessToken.startsWith(ACCESS_TOKEN_STARTS)) {
-            return accessToken.substring(BEGIN_INDEX);
+    private String resolveToken(String refreshToken) {
+        jwtTokenProvider.validateToken(refreshToken);
+        if (refreshToken.startsWith(ACCESS_TOKEN_STARTS)) {
+            return refreshToken.substring(BEGIN_INDEX);
         }
         throw new CannotResolveToken(REFRESH_TOKEN_RESOLVE, REFRESH_TOKEN_RESOLVE.getMessage());
     }
